Task4: Products filter - Feature: Adding product filter onto a seperate component
